compass/apps/api/pyproject.toml
[project]
name = "compass-api"
version = "0.1.0"
description = "Compass API — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "redis[hiredis]>=5.2.0",
    "arq>=0.26.0",
    "anthropic>=0.40.0",
    "jinja2>=3.1.4",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "python-multipart>=0.0.12",
    "pgvector>=0.3.6",
    "sentence-transformers>=3.3.0",
    "langgraph>=0.2.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
    "pytest-mock>=3.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
